Added lines of code that uses the library.
I'm still not sure what to put in the ESP32PWM::allocateTimer() part,
made the variable just a constexpr so that i may be able to change it
later down the line. perhaps they use the same timer? i don't really
know lmao

Also using writeMicroseconds because the servos go from 0-270 deg, which
isn't the 0-180 deg standard.

+++ include/main.hh
@@ -51,11 +51,11 @@ typedef struct {
 typedef struct {
     int16_t left_motors;
     int16_t right_motors;
-    uint8_t servo_out1;
-    uint8_t servo_out2;
-    uint8_t servo_out3;
-    uint8_t servo_out4;
-}motors_t;
+    uint16_t servo_out1;
+    uint16_t servo_out2;
+    uint16_t servo_out3;
+    uint16_t servo_out4;
+} motors_t;
 
 
 // Pin names
